Refactor structs package as a standard go package practice
