+++ utils/structs/structs.go
@@ -1,11 +1,11 @@
-package utils
+package structs
 
 import (
 	"errors"
 	"reflect"
 )
 
-//ArrayValues returns strut values as an array
+// ArrayValues returns strut values as an array
 func ArrayValues(arr interface{}) []interface{} {
 	v := reflect.ValueOf(arr)
 	values := make([]interface{}, v.NumField())
@@ -15,7 +15,7 @@ func ArrayValues(arr interface{}) []interface{} {
 	return values
 }
 
-//ArrayContains checks if exists in array
+// ArrayContains checks if exists in array
 func ArrayContains(arr []interface{}, str string) bool {
 	for _, v := range arr {
 		if v == str {
@@ -25,7 +25,7 @@ func ArrayContains(arr []interface{}, str string) bool {
 	return false
 }
 
-//ExistsInStrut if a value exists returns true
+// ExistsInStrut if a value exists returns true
 func ExistsInStrut(arr interface{}, toCheck string) bool {
 	v := reflect.ValueOf(arr)
 	for i := 0; i < v.NumField(); i++ {
@@ -36,7 +36,7 @@ func ExistsInStrut(arr interface{}, toCheck string) bool {
 	return false
 }
 
-//GetStructFieldByString returns the named field as an interface{}, also returns the Type of the interface.
+// GetStructFieldByString returns the named field as an interface{}, also returns the Type of the interface.
 func GetStructFieldByString(arr interface{}, toGet string) (interface{}, string, error) {
 	v := reflect.ValueOf(arr)
 	t := v.Kind().String()
@@ -49,6 +49,5 @@ func GetStructFieldByString(arr interface{}, toGet string) (interface{}, string,
 		err := errors.New("GetStrutFieldByString(): cannot find field in struct")
 		return nil, "", err
 	}
-	//return f.Interface(), f.Type().String(), nil
 	return f.Interface(), f.Kind().String(), nil
 }
